brytit ut produkterna från layouten
